Moved request dto to requestMapper
