+++ src/main/java/com/salessparrow/api/lib/crmActions/createNote/CreateNoteInterface.java
@@ -3,8 +3,8 @@ package com.salessparrow.api.lib.crmActions.createNote;
 import org.springframework.stereotype.Component;
 
 import com.salessparrow.api.domain.SalesforceUser;
-import com.salessparrow.api.dto.NoteDto;
 import com.salessparrow.api.dto.formatter.CreateNoteFormatterDto;
+import com.salessparrow.api.dto.requestMapper.NoteDto;
 
 /**
  * CreateNoteInterface is an interface for the create note action for the CRM.
