+++ src/main/java/com/salessparrow/api/lib/crmActions/createNote/CreateSalesforceNote.java
@@ -14,9 +14,9 @@ import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.salessparrow.api.domain.SalesforceUser;
-import com.salessparrow.api.dto.NoteDto;
 import com.salessparrow.api.dto.formatter.CreateNoteFormatterDto;
 import com.salessparrow.api.exception.CustomException;
+import com.salessparrow.api.dto.requestMapper.NoteDto;
 import com.salessparrow.api.lib.Base64Helper;
 import com.salessparrow.api.lib.Util;
 import com.salessparrow.api.lib.errorLib.ErrorObject;
