+++ src/main/java/com/salessparrow/api/services/salesforce/AuthService.java
@@ -10,8 +10,8 @@ import com.salessparrow.api.config.CoreConstants;
 import com.salessparrow.api.domain.SalesforceOauthToken;
 import com.salessparrow.api.domain.SalesforceOrganization;
 import com.salessparrow.api.domain.SalesforceUser;
-import com.salessparrow.api.dto.SalesforceConnectDto;
 import com.salessparrow.api.dto.entities.CurrentUserEntityDto;
+import com.salessparrow.api.dto.requestMapper.SalesforceConnectDto;
 import com.salessparrow.api.exception.CustomException;
 import com.salessparrow.api.lib.AwsKms;
 import com.salessparrow.api.lib.CookieHelper;
